requirements.txt
pandas==1.3.3
numpy==1.21.2
psycopg2-binary==2.9.1
sqlalchemy==1.4.23
tqdm==4.66.5